feat: extend mongo property data with subgraph data

+++ src/services/property.service.ts
@@ -1,31 +1,129 @@
+// Exceptions
 import { HttpException } from "@exceptions/HttpException";
-import { Property } from "@interfaces/property.interface";
-import { IPaginateResult } from "typegoose-cursor-pagination";
+
+// models
 import propertyModel from "@models/property.model";
-import { isEmpty, parseOptions } from "@utils/util";
-import type { Options } from "@utils/types";
+import iroModel from "@models/iro.model";
+
+// interfaces
+import { Property, PropertyExtended } from "@interfaces/property.interface";
+import { IROReduced } from "@interfaces/iro.interface";
+import { FilterQuery, PaginateOptions } from 'mongoose';
+
+// utils
+import { isEmpty } from "@utils/util";
+import BigNumber from "bignumber.js";
+
+// configure BigNumber constructor
+BigNumber.config({ DECIMAL_PLACES: 2 });
+
+// instantiate models
+const iro = new iroModel();
+
+interface GetPropertiesPaginatedResult {
+  docs: PropertyExtended[]
+  metadata: {
+    totalDocs: number;
+    limit: number;
+    hasPrevPage: boolean;
+    hasNextPage: boolean;
+    page?: number | undefined;
+    totalPages: number;
+    offset: number;
+    prevPage?: number | null | undefined;
+    nextPage?: number | null | undefined;
+    pagingCounter: number;
+  }
+}
 
 class PropertyService {
   // TODO: Implement DTO for propertyBody
   public async createProperty(propertyBody): Promise<Property> {
     if (isEmpty(propertyBody)) throw new HttpException(400, "propertyBody is empty");
 
+    const findOne = await propertyModel.findOne({name: propertyBody.name });
+    if (findOne) throw new HttpException(409, "Property already exist (by name)");
+
     return await propertyModel.create(propertyBody);
   }
 
-  // TODO: Translate filter to mongoDB format
-  public async getPropertiesPaginated(filter, options: Options): Promise<IPaginateResult<Property>> {
-    const propertiesPagination = await propertyModel.findPaged(parseOptions(options));
-    return propertiesPagination as IPaginateResult<Property>;
+  public async getPropertiesPaginated(filter: FilterQuery<Property>, options: PaginateOptions): Promise<GetPropertiesPaginatedResult> {
+    const propertiesPagination = await propertyModel.paginate(filter, options);
+
+    const iroIds: string[] = [];
+    propertiesPagination.docs.forEach(doc => {
+      if (doc.status === "crowdfunding") {
+        iroIds.push(doc.iroId.toString());
+      }
+    });
+
+    const results: GetPropertiesPaginatedResult = {
+      metadata: {
+        totalDocs: propertiesPagination.totalDocs,
+        limit: propertiesPagination.limit,
+        hasPrevPage: propertiesPagination.hasPrevPage,
+        hasNextPage: propertiesPagination.hasNextPage,
+        page: propertiesPagination.page, 
+        totalPages: propertiesPagination.totalPages,
+        offset: propertiesPagination.offset,
+        prevPage: propertiesPagination.prevPage,
+        nextPage: propertiesPagination.nextPage, 
+        pagingCounter: propertiesPagination.pagingCounter
+      },
+      docs: []
+    };
+    const iroQueryResult = await iro.getIros(iroIds);
+    const iros: {[iroId: string]: IROReduced } = {};
+    iroQueryResult.iros.forEach(iro => {
+      iros[iro.iroId] = iro;
+    });
+
+    propertiesPagination.docs.forEach(doc => {
+      const property: PropertyExtended = {...doc};
+      results.docs.push(property);
+      if (doc.status === "crowdfunding") {
+        const iro = iros[doc.iroId];
+        const denominator = new BigNumber(iro.currencyDecimals);
+        property.iroStatus = iro.status;
+        property.iroUnitPrice = (new BigNumber(iro.unitPrice))
+          .div(denominator).toString();
+        property.iroCurrency = iro.currency;
+        property.iroSoftCap = (new BigNumber(iro.softCap))
+          .div(denominator).toString();
+        property.iroHardCap = (new BigNumber(iro.hardCap))
+          .div(denominator).toString();
+        property.iroStart = iro.start;
+        property.iroEnd = iro.end;
+      }
+    });
+
+    return results;
   }
 
-  public async getPropertyById(propertyId: string): Promise<Property> {
+  public async getPropertyById(propertyId: string): Promise<PropertyExtended> {
     if (isEmpty(propertyId)) throw new HttpException(400, "propertyId is empty");
 
     const property = await propertyModel.findById(propertyId);
     if (!property) throw new HttpException(409, "Property doesn't exist");
 
-    return property;
+    const result: PropertyExtended = {...property};
+    if (property.status === "crowdfunding") {
+      const iroQueryResult = (await iro.getIros([property.iroId.toString()]))
+        .iros[0];
+      const denominator = new BigNumber(iroQueryResult.currencyDecimals);
+      result.iroStatus = iroQueryResult.status;
+      result.iroUnitPrice = (new BigNumber(iroQueryResult.unitPrice))
+        .div(denominator).toString();
+      result.iroCurrency = iroQueryResult.currency;
+      result.iroSoftCap = (new BigNumber(iroQueryResult.softCap))
+        .div(denominator).toString();
+      result.iroHardCap = (new BigNumber(iroQueryResult.hardCap))
+        .div(denominator).toString();
+      result.iroStart = iroQueryResult.start;
+      result.iroEnd = iroQueryResult.end;
+    }
+
+    return result;
   }
 
   public async getPropertyByName(name: string): Promise<Property> {
@@ -39,12 +137,16 @@ class PropertyService {
 
   // TODO: Implement some checks and DTO for updateBody param
   public async updatePropertyById(propertyId: string, updateBody): Promise<Property> {
+    if (isEmpty(propertyId)) throw new HttpException(400, "propertyId is empty");
+    if (isEmpty(updateBody)) throw new HttpException(400, "updateBody is empty");
     const updatedProperty = propertyModel.findByIdAndUpdate(propertyId, updateBody);
 
     return updatedProperty;
   }
 
   public async setIroId(propertyId: string, iroId: string) {
+    if (isEmpty(propertyId)) throw new HttpException(400, "propertyId is empty");
+    if (isEmpty(iroId)) throw new HttpException(400, "iroId is empty");
     const updatedProperty = await propertyModel.findByIdAndUpdate(propertyId, { 
       iroId,
       stage: "crowdfunding"
@@ -53,6 +155,8 @@ class PropertyService {
   }
 
   public async setRealEstateNftId(propertyId: string, realEstateNftId: string) {
+    if (isEmpty(propertyId)) throw new HttpException(400, "propertyId is empty");
+    if (isEmpty(realEstateNftId)) throw new HttpException(400, "realEstateNftId is empty");
     const updatedProperty = await propertyModel.findByIdAndUpdate(propertyId, { 
       realEstateNftId,
       stage: "trade"
@@ -61,6 +165,7 @@ class PropertyService {
   }
 
   public async deletePropertyById(propertyId: string): Promise<Property> {
+    if (isEmpty(propertyId)) throw new HttpException(400, "propertyId is empty");
     const removedProperty = propertyModel.findByIdAndRemove(propertyId);
     return removedProperty;
   }
